It now checks the View in BRAND link for action notification emails
It now needs a new keys in config file to get be able to login to the email that receives the notifications emails.

It will find the email by subject and find the "View in BRAND link" and load that link's URL and see if it matches the expected format.

#5